backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.7.1
flask-cors==4.0.1
werkzeug==3.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyjwt==2.8.0
openai>=1.51.0